# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: anthropic
    description: Anthropic AI chat operations
  - name: study
    description: Study material and learning operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /anthropic/conversations:
    get:
      operationId: listAnthropicConversations
      tags: [anthropic]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicConversation"
    post:
      operationId: createAnthropicConversation
      tags: [anthropic]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnthropicConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"
  /anthropic/conversations/{id}:
    get:
      operationId: getAnthropicConversation
      tags: [anthropic]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteAnthropicConversation
      tags: [anthropic]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
  /anthropic/conversations/{id}/messages:
    get:
      operationId: listAnthropicMessages
      tags: [anthropic]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicMessage"
    post:
      operationId: sendAnthropicMessage
      tags: [anthropic]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAnthropicMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /study/generate-quiz:
    post:
      operationId: generateQuiz
      tags: [study]
      summary: Generate quiz from study material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStudyContentBody"
      responses:
        "200":
          description: Generated quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizQuestion"
  /study/generate-flashcards:
    post:
      operationId: generateFlashcards
      tags: [study]
      summary: Generate flashcards from study material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStudyContentBody"
      responses:
        "200":
          description: Generated flashcards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"
  /study/generate-summary:
    post:
      operationId: generateSummary
      tags: [study]
      summary: Generate summary from study material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStudyContentBody"
      responses:
        "200":
          description: Generated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySummary"
  /study/gap-detection:
    post:
      operationId: detectGaps
      tags: [study]
      summary: Detect knowledge gaps from quiz results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GapDetectionBody"
      responses:
        "200":
          description: Detected knowledge gaps and recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GapDetectionResult"
  /study/study-plan:
    post:
      operationId: generateStudyPlan
      tags: [study]
      summary: Generate a personalized study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanBody"
      responses:
        "200":
          description: Generated study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
  /study/sessions:
    get:
      operationId: listStudySessions
      tags: [study]
      summary: List all study sessions
      responses:
        "200":
          description: List of study sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudySession"
    post:
      operationId: createStudySession
      tags: [study]
      summary: Create/save a study session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudySessionBody"
      responses:
        "201":
          description: Created study session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"
  /study/progress:
    get:
      operationId: getStudyProgress
      tags: [study]
      summary: Get overall study progress and analytics
      responses:
        "200":
          description: Study progress data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyProgress"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateAnthropicConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendAnthropicMessageBody:
      type: object
      properties:
        content:
          type: string
        systemPrompt:
          type: string
      required:
        - content
    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    GenerateStudyContentBody:
      type: object
      properties:
        material:
          type: string
        topic:
          type: string
        count:
          type: integer
      required:
        - material
    QuizQuestion:
      type: object
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        answer:
          type: integer
        explanation:
          type: string
        difficulty:
          type: string
      required:
        - question
        - options
        - answer
        - explanation
    Flashcard:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
        hint:
          type: string
      required:
        - front
        - back
    StudySummary:
      type: object
      properties:
        title:
          type: string
        overview:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        concepts:
          type: array
          items:
            type: object
            properties:
              term:
                type: string
              definition:
                type: string
            required:
              - term
              - definition
        examTips:
          type: array
          items:
            type: string
        fullText:
          type: string
      required:
        - title
        - overview
        - keyPoints
        - fullText
    GapDetectionBody:
      type: object
      properties:
        material:
          type: string
        wrongAnswers:
          type: array
          items:
            type: object
            properties:
              question:
                type: string
              yourAnswer:
                type: string
              correctAnswer:
                type: string
            required:
              - question
              - yourAnswer
              - correctAnswer
      required:
        - material
        - wrongAnswers
    GapDetectionResult:
      type: object
      properties:
        weakAreas:
          type: array
          items:
            type: object
            properties:
              topic:
                type: string
              severity:
                type: string
              recommendation:
                type: string
            required:
              - topic
              - severity
              - recommendation
        overallScore:
          type: number
        studyPriorities:
          type: array
          items:
            type: string
      required:
        - weakAreas
        - overallScore
        - studyPriorities
    StudyPlanBody:
      type: object
      properties:
        material:
          type: string
        goal:
          type: string
        daysAvailable:
          type: integer
      required:
        - material
    StudyPlan:
      type: object
      properties:
        title:
          type: string
        totalDays:
          type: integer
        sessions:
          type: array
          items:
            type: object
            properties:
              day:
                type: integer
              focus:
                type: string
              activities:
                type: array
                items:
                  type: string
              duration:
                type: string
            required:
              - day
              - focus
              - activities
              - duration
      required:
        - title
        - totalDays
        - sessions
    StudySession:
      type: object
      properties:
        id:
          type: integer
        topic:
          type: string
        mode:
          type: string
        score:
          type: number
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        durationMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - topic
        - mode
        - createdAt
    CreateStudySessionBody:
      type: object
      properties:
        topic:
          type: string
        mode:
          type: string
        score:
          type: number
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        durationMinutes:
          type: integer
      required:
        - topic
        - mode
    StudyProgress:
      type: object
      properties:
        totalSessions:
          type: integer
        totalMinutes:
          type: integer
        averageScore:
          type: number
        streakDays:
          type: integer
        recentSessions:
          type: array
          items:
            $ref: "#/components/schemas/StudySession"
        topTopics:
          type: array
          items:
            type: object
            properties:
              topic:
                type: string
              sessions:
                type: integer
              avgScore:
                type: number
            required:
              - topic
              - sessions
              - avgScore
      required:
        - totalSessions
        - totalMinutes
        - averageScore
        - streakDays
        - recentSessions
        - topTopics